Update inventory.md
According to https://github.com/ansible-semaphore/semaphore/issues/47#issuecomment-376561708 using a file should work, but I could not find it in the docs. After taking a leap of faith and testing on my install, it seems to work.

If you run a play that is located in a git repo, then Semaphore will actually search for the inventory file in the location specified, but relative to your playbook location. So using standard Ansible practices, creating a folder in your git repo called `inventory` and then putting your yaml file in there, will actually work when referenced in the "inventory" configuration in Semaphore UI.